+++ TSE/SFXBolt.cpp
@@ -65,8 +65,8 @@ void CBoltEffectCreator::Paint (CG32bitImage &Dest, int x, int y, SViewportPaint
 		xStart = x - (int)(vTail.GetX() + 0.5);
 		yStart = y + (int)(vTail.GetY() + 0.5);
 
-		rgbStart = CG32bitPixel(m_rgbSecondaryColor, (BYTE)155);
-		rgbEnd = CG32bitPixel(m_rgbSecondaryColor, (BYTE)0);
+		rgbStart = CG32bitPixel(m_rgbSecondaryColor, 155);
+		rgbEnd = CG32bitPixel(m_rgbSecondaryColor, 0);
 		CGDraw::LineGradient(Dest, xStart, yStart, x, y, m_iWidth, rgbEnd, rgbStart);
 		}
 
@@ -75,7 +75,7 @@ void CBoltEffectCreator::Paint (CG32bitImage &Dest, int x, int y, SViewportPaint
 	yStart = y + (int)(vTail.GetY() + 0.5);
 
 	rgbStart = m_rgbPrimaryColor;
-	rgbEnd = CG32bitPixel(m_rgbSecondaryColor, (BYTE)200);
+	rgbEnd = CG32bitPixel(m_rgbSecondaryColor, 200);
 	CGDraw::LineGradient(Dest, xStart, yStart, x, y, Max(1, m_iWidth / 2), rgbEnd, rgbStart);
 	}
